v0.17.1.0: fix bug in constructEquivClasses
This commit fixes a bug in constructEquivClasses in which we failed to
use equivalences in both directions.
It also refactors that function to use IntMaps rather than Maps.
For the case we were working with, TyVars, we were already comparing
based on the Uniques, so we might as well use the appropriate data
structures.